chore: normalize directories of activities and fragments
