+++ app/app/src/main/java/com/architecture/app/screens/MainActivity.java
@@ -1,4 +1,4 @@
-package com.architecture.app;
+package com.architecture.app.screens;
 
 import android.content.Intent;
 import android.os.Build;
@@ -8,8 +8,8 @@ import android.widget.Button;
 import androidx.annotation.RequiresApi;
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.architecture.app.R;
 import com.architecture.app.components.Navbar;
-import com.architecture.app.screens.UploadActivity;
 
 public class MainActivity extends AppCompatActivity {
     private Button _uploadButton;
