+++ app/app/src/main/java/com/architecture/app/screens/UploadActivity.java
@@ -58,8 +58,10 @@ public class UploadActivity extends AppCompatActivity {
     }
 
     @Override
-    public boolean onOptionsItemSelected(MenuItem item){
-        finish();
+    public boolean onOptionsItemSelected(MenuItem item) {
+        if(item.getItemId() == R.id.home) {
+            finish();
+        }
         return true;
     }
