+++ app/app/src/main/res/layout/fragment_check.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/purple_light"
-    tools:context=".fragments.CheckFragment">
+    tools:context=".screens.fragments.CheckFragment">
 
     <TextView
         android:layout_width="wrap_content"
