+++ app/app/src/main/res/layout/fragment_home.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/gray_light"
-    tools:context=".fragments.HomeFragment">
+    tools:context=".screens.fragments.HomeFragment">
 
     <TextView
         android:layout_width="wrap_content"
