+++ app/app/src/main/res/layout/fragment_question.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/yellow"
-    tools:context=".fragments.QuestionFragment">
+    tools:context=".screens.fragments.QuestionFragment">
 
     <TextView
         android:layout_width="wrap_content"
